Move analysis into activities
