+++ oonipipeline/src/oonipipeline/temporal/workflows.py
@@ -3,7 +3,15 @@ import logging
 from typing import List
 from oonidata.dataclient import date_interval
 from oonidata.datautils import PerfTimer
-from oonipipeline.temporal.common import optimize_all_tables
+from oonipipeline.db.connections import ClickhouseConnection
+from oonipipeline.temporal.activities.analysis import (
+    AnalysisWorkflowParams,
+    MakeAnalysisParams,
+    log,
+    make_analysis_in_a_day,
+    make_cc_batches,
+)
+from oonipipeline.temporal.common import get_obs_count_by_cc, optimize_all_tables
 from oonipipeline.temporal.activities.observations import (
     MakeObservationsParams,
     make_observation_in_day,
@@ -141,3 +149,82 @@ class GroundTruthsWorkflow:
                     )
                 )
                 task_list.append(task)
+
+
+@workflow.defn(sandboxed=False)
+class AnalysisWorkflow:
+    @workflow.run
+    async def run(self, params: AnalysisWorkflowParams) -> dict:
+        t_total = PerfTimer()
+
+        t = PerfTimer()
+        start_day = datetime.strptime(params.start_day, "%Y-%m-%d").date()
+        end_day = datetime.strptime(params.end_day, "%Y-%m-%d").date()
+
+        log.info("building ground truth databases")
+
+        async with asyncio.TaskGroup() as tg:
+            for day in date_interval(start_day, end_day):
+                tg.create_task(
+                    workflow.execute_activity(
+                        make_ground_truths_in_day,
+                        MakeGroundTruthsParams(
+                            day=day.strftime("%Y-%m-%d"),
+                            clickhouse=params.clickhouse,
+                            data_dir=params.data_dir,
+                            rebuild_ground_truths=params.rebuild_ground_truths,
+                        ),
+                        start_to_close_timeout=timedelta(minutes=2),
+                    )
+                )
+            log.info(f"built ground truth db in {t.pretty}")
+
+        with ClickhouseConnection(params.clickhouse) as db:
+            cnt_by_cc = get_obs_count_by_cc(
+                db, start_day=start_day, end_day=end_day, test_name=params.test_name
+            )
+        cc_batches = make_cc_batches(
+            cnt_by_cc=cnt_by_cc,
+            probe_cc=params.probe_cc,
+            parallelism=params.parallelism,
+        )
+        log.info(
+            f"starting processing of {len(cc_batches)} batches over {(end_day - start_day).days} days (parallelism = {params.parallelism})"
+        )
+        log.info(f"({cc_batches} from {start_day} to {end_day}")
+
+        task_list = []
+        async with asyncio.TaskGroup() as tg:
+            for probe_cc in cc_batches:
+                for day in date_interval(start_day, end_day):
+                    task = tg.create_task(
+                        workflow.execute_activity(
+                            make_analysis_in_a_day,
+                            MakeAnalysisParams(
+                                probe_cc=probe_cc,
+                                test_name=params.test_name,
+                                clickhouse=params.clickhouse,
+                                data_dir=params.data_dir,
+                                fast_fail=params.fast_fail,
+                                day=day.strftime("%Y-%m-%d"),
+                            ),
+                            start_to_close_timeout=timedelta(minutes=30),
+                        )
+                    )
+                    task_list.append(task)
+
+        t = PerfTimer()
+        # size, msmt_count =
+        total_obs_count = 0
+        for task in task_list:
+            res = task.result()
+
+            total_obs_count += res["count"]
+
+        log.info(f"produces a total of {total_obs_count} analysis")
+        obs_per_sec = round(total_obs_count / t_total.s)
+        log.info(
+            f"finished processing {start_day} - {end_day} speed: {obs_per_sec}obs/s)"
+        )
+        log.info(f"{total_obs_count} msmts in {t_total.pretty}")
+        return {"total_obs_count": total_obs_count}
